fixing more php 8 phpstan errors

+++ lib/Doctrine/Locking/Manager/Pessimistic.php
@@ -120,6 +120,10 @@ class Doctrine_Locking_Manager_Pessimistic
                                   . ' (object_type, object_key, user_ident, timestamp_obtained)'
                                   . ' VALUES (:object_type, :object_key, :user_ident, :ts_obtained)');
 
+            if (!$stmt) {
+                throw new Doctrine_Locking_Exception('Could not prepare database statement for lock');
+            }
+
             $stmt->bindParam(':object_type', $objectType);
             $stmt->bindParam(':object_key', $key);
             $stmt->bindParam(':user_ident', $userIdent);
@@ -144,6 +148,9 @@ class Doctrine_Locking_Manager_Pessimistic
                                           . ' WHERE object_type = :object_type AND'
                                           . ' object_key  = :object_key  AND'
                                           . ' user_ident  = :user_ident');
+                    if (!$stmt) {
+                        throw new Doctrine_Locking_Exception('Could not prepare database statement for lock');
+                    }
                     $stmt->bindParam(':ts', $time);
                     $stmt->bindParam(':object_type', $objectType);
                     $stmt->bindParam(':object_key', $key);
@@ -184,6 +191,11 @@ class Doctrine_Locking_Manager_Pessimistic
                                         object_type = :object_type AND
                                         object_key  = :object_key  AND
                                         user_ident  = :user_ident");
+
+            if (!$stmt) {
+                throw new Doctrine_Locking_Exception('Could not prepare database statement for lock');
+            }
+
             $stmt->bindParam(':object_type', $objectType);
             $stmt->bindParam(':object_key', $key);
             $stmt->bindParam(':user_ident', $userIdent);
@@ -218,6 +230,11 @@ class Doctrine_Locking_Manager_Pessimistic
             $dbh  = $this->conn->getDbh();
             $stmt = $dbh->prepare('SELECT user_ident FROM ' . $this->_lockTable
                                   . ' WHERE object_type = :object_type AND object_key = :object_key');
+
+            if (!$stmt) {
+                throw new Doctrine_Locking_Exception('Could not prepare database statement for lock');
+            }
+
             $stmt->bindParam(':object_type', $objectType);
             $stmt->bindParam(':object_key', $key);
             $success = $stmt->execute();
@@ -266,6 +283,11 @@ class Doctrine_Locking_Manager_Pessimistic
         try {
             $dbh  = $this->conn->getDbh();
             $stmt = $dbh->prepare('DELETE FROM ' . $this->_lockTable . ' WHERE timestamp_obtained < :age');
+
+            if (!$stmt) {
+                throw new Doctrine_Locking_Exception('Could not prepare database statement for lock');
+            }
+
             $stmt->bindParam(':age', $age);
             $query = 'DELETE FROM ' . $this->_lockTable . ' WHERE timestamp_obtained < :age';
             if ($objectType) {
@@ -275,6 +297,11 @@ class Doctrine_Locking_Manager_Pessimistic
                 $query .= ' AND user_ident = :user_ident';
             }
             $stmt = $dbh->prepare($query);
+
+            if (!$stmt) {
+                throw new Doctrine_Locking_Exception('Could not prepare database statement for lock');
+            }
+
             $stmt->bindParam(':age', $age);
             if ($objectType) {
                 $stmt->bindParam(':object_type', $objectType);
